:bug: fix: ajusta path e atributo
Ajustes de path em contatos e atributo acompanhante para compatibilidade com banco

+++ src/main/java/br/com/fiap/dto/AcessoRequestDTO.java
@@ -5,8 +5,8 @@ import java.util.UUID;
 
 public record AcessoRequestDTO(
     String funcionalidade,
-    int quantidadeAcessos,
-    int tempoPermanenciaSeg,
+    float quantidadeAcessos,
+    float tempoPermanenciaSeg,
     LocalDateTime dataAcesso,
     UUID idPaciente
 ) {}
